combine shuffle test to initial test
